Combine the ModelTags concern with Machine
Originally the ModelTags concern was going to be reused in a few
different models, however these requirement has since been removed.

It should therefore be merged with the `Machine` model has this is the
only time it gets used
