+++ lib/cloudware/models/machine.rb
@@ -30,14 +30,57 @@ require 'cloudware/models/concerns/provider_client'
 module Cloudware
   module Models
     class Machine < Application
+      class << self
+        def flag
+          'TAG'
+        end
+
+        def build_from_context(context)
+          (context.results || {})
+            .keys
+            .map { |k| name_from_tag(k) }
+            .uniq
+            .reject(&:nil?)
+            .map do |name|
+            new(name: name, context: context)
+          end
+        end
+
+        #
+        # Extracts the models name from a tag
+        #
+        # Valid tags are given in the format:
+        # <name>TAG<tag>
+        #
+        # The regex matches the '<name>' component of the above tag if the
+        # syntax is correct, otherwise it returns nil.
+        #
+        # It therefore can also be used to filter out arbitrary tags from
+        # model tags
+        #
+        # Examples:
+        #   name_from_tag('node01TAGsome-key') => 'node01'
+        #   name_from_tag('some-other-string') =>  nil
+        #
+        def name_from_tag(tag)
+          regex = /\A.*(?=#{flag}.*\Z)/
+          regex.match(tag.to_s)&.to_a&.first
+        end
+
+        def tag_generator(name, tag)
+          :"#{name}#{flag}#{tag}"
+        end
+      end
+
       include Concerns::ProviderClient
-      include Concerns::ModelTags
 
       PROVIDER_ID_FLAG = 'ID'
       GROUPS_TAG = 'groups'
 
       delegate :status, :off, :on, to: :machine_client
       delegate :region, :provider, to: :context
+
+      attr_accessor :name, :context
 
       def provider_id
         fetch_result(PROVIDER_ID_FLAG) do |long_tag|
@@ -50,6 +93,45 @@ module Cloudware
 
       def groups
         fetch_result(GROUPS_TAG, default: '').split(',')
+      end
+
+      def tags
+        (context.results || {}).each_with_object({}) do |(key, value), memo|
+          next unless (tag = extract_tag(key))
+          memo[tag] = value
+        end
+      end
+
+      def tag_generator(tag)
+        self.class.tag_generator(name, tag)
+      end
+
+      #
+      # Extract the tag component from the key
+      #
+      # Valid tags are given in the format:
+      # <name>TAG<tag>
+      #
+      # The regex matches the `<tag>` component of the above tag if the
+      # syntax matches with the correct name. Otherwise it returns `nil`
+      #
+      # Examples:
+      #
+      #   extract_tag('mynodeTAGkey')        => 'key'
+      #   extract_tag('differentnodeTAGkey') =>  nil
+      #   extract_tag('random-string')       =>  nil
+      #
+      def extract_tag(key)
+        regex = /(?<=\A#{self.class.tag_generator(Regexp.escape(name), '')}).*/
+        regex.match(key.to_s)&.to_a&.first&.to_sym
+      end
+
+      def fetch_result(short_tag, default: nil)
+        long_tag = tag_generator(short_tag)
+        result = (context.results || {})[long_tag]
+        return result unless result.nil?
+        return default unless default.nil?
+        yield long_tag if block_given?
       end
 
       private
